fix: prevent name conflict between model and factory fixtures
On factory registration there was a possibility to specify name of the model
fixture the same as the name of the factory fixture created by
lowercase-underscore convention. This would lead, depending on pytest version to
`recursive dependency` or `fixture not found` error, which would be confusing
to the user. To prevent that scenario from happening, additional assertion was
created, which would detect such cases and provide explicit error message.